% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{country_table}
\alias{country_table}
\title{Create table with just countries and their codes}
\usage{
country_table(include_codes = TRUE, include_names = TRUE,
  stringsAsFactors = FALSE)
}
\value{
A data frame with country codes, names, or both. Always a
    data frame, even if one of \code{include_codes} or \code{include_names} is
    \code{FALSE}.
}
\description{
Creates a table with country codes and their names. Optionally,
just a table with names or codes. This function returns a data
frame. If only one of \code{include_codes} and \code{include_names} is \code{TRUE}
it will be a data frame with one column. If you want a vector of
names or codes see \code{\link{code}} and \code{\link{name}}.
}
\seealso{
Other Table functions: \code{\link{reg_table}}
}
\author{
Mark Wheldon
}
\concept{Table functions}
