% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_codes}
\alias{list_codes}
\title{List all codes of countries, regions, etc}
\usage{
list_codes()
}
\value{
Numeric vector of unique codes (countries, regions, etc)
}
\description{
Lists the unique codes in \code{unloc_df$code}. \code{unloc_df} is
from the \pkg{wpp2019} package.
}
\seealso{
Other listing functions: 
\code{\link{list_country_codes}()},
\code{\link{list_country_names}()},
\code{\link{list_names}()},
\code{\link{list_reg_codes}()},
\code{\link{list_reg_names}()},
\code{\link{list_sdg_colours}()}
}
\author{
Mark Wheldon
}
\concept{listing functions}
