% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_country_codes}
\alias{list_country_codes}
\title{List country codes}
\usage{
list_country_codes()
}
\value{
Numeric vector of country codes.
}
\description{
All country codes are listed.
}
\seealso{
Other listing functions: 
\code{\link{list_codes}()},
\code{\link{list_country_names}()},
\code{\link{list_names}()},
\code{\link{list_reg_codes}()},
\code{\link{list_reg_names}()},
\code{\link{list_sdg_colours}()}
}
\author{
Mark Wheldon
}
\concept{listing functions}
