% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_country_names}
\alias{list_country_names}
\title{List all country names}
\usage{
list_country_names()
}
\value{
Character vector of country names.
}
\description{
All country names are listed.
}
\seealso{
Other listing functions: \code{\link{list_codes}},
  \code{\link{list_country_codes}},
  \code{\link{list_names}}, \code{\link{list_reg_codes}},
  \code{\link{list_reg_names}}
}
\author{
Mark Wheldon
}
\concept{listing functions}
