% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_reg_colours}
\alias{list_reg_colours}
\alias{list_reg_colors}
\title{List region colours}
\usage{
list_reg_colours(
  family = c("M49", "SDG", "WB_Inc", "Dev"),
  model = c("hex", "rgb"),
  use_reg_names = FALSE,
  add_world = FALSE
)

list_reg_colors(
  family = c("M49", "SDG", "WB_Inc", "Dev"),
  model = c("hex", "rgb"),
  use_reg_names = FALSE,
  add_world = FALSE
)
}
\arguments{
\item{family}{Region family for which to return colour codes.}

\item{model}{Colour model (only \code{"hex"} and \code{"rgb"} are
supported).}

\item{use_reg_names}{Logial. Should the names of the output object
be the names of the region? Otherwise they are the codes.}

\item{add_world}{Logical. Should \dQuote{World} be added to the
output?}
}
\value{
Named character vector or list of region colours with
    region codes as names (or region names if \code{use_reg_names}
    is \code{TRUE}).
}
\description{
List hex or RGB codes of the official colours to be used for
plotting various level 1 regions (there are no colours for level 2
regions).
}
\examples{
list_reg_colours("SDG", "hex")
list_reg_colors("SDG", "hex")
}
\seealso{
Other listing functions: 
\code{\link{list_codes}()},
\code{\link{list_country_codes}()},
\code{\link{list_country_names}()},
\code{\link{list_names}()},
\code{\link{list_reg_codes}()},
\code{\link{list_reg_names}()}
}
\author{
Mark Wheldon
}
\concept{listing functions}
