% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_reg_names}
\alias{list_reg_names}
\title{List region names}
\usage{
list_reg_names(level = c("1", "2"), family = c("M49", "SDG", "WB_inc", "Dev"))
}
\arguments{
\item{level}{Level of region. Higher levels are nested in
lower levels. E.g., \dQuote{Africa} is level \dQuote{1},
\dQuote{Eastern Africa} is level \dQuote{2}. Converted to
character if supplied as numeric.}

\item{family}{Family to which the region referenced by \code{code}
belongs.}
}
\value{
Character vector of region names
}
\description{
List all region names for the given family and level.
}
\seealso{
Other listing functions: 
\code{\link{list_codes}()},
\code{\link{list_country_codes}()},
\code{\link{list_country_names}()},
\code{\link{list_names}()},
\code{\link{list_reg_codes}()}
}
\author{
Mark Wheldon
}
\concept{listing functions}
