% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_sdg_colours}
\alias{list_sdg_colours}
\alias{list_sdg_colors}
\title{List SDG region colours}
\usage{
list_sdg_colours(model = c("hex", "rgb"))

list_sdg_colors(model = c("hex", "rgb"))
}
\arguments{
\item{model}{Colour model (only 'hex' and 'rgb' are supported).}
}
\value{
Character vector or list of SDG colours
}
\description{
List hex or RGB codes for the official colours to be used for plotting SDG regions.
}
\examples{
list_sdg_colours("hex")
list_sdg_colors("hex")
}
\seealso{
Other listing functions: 
\code{\link{list_codes}()},
\code{\link{list_country_codes}()},
\code{\link{list_country_names}()},
\code{\link{list_names}()},
\code{\link{list_reg_codes}()},
\code{\link{list_reg_names}()}
}
\author{
Mark Wheldon
}
\concept{listing functions}
