% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{loc_table}
\alias{loc_table}
\title{.. title (sentence case, not terminated with a full-stop) ..}
\usage{
loc_table(level = c("1", "2"), family = c("M49", "SDG", "WB", "Dev"),
  codes = TRUE, names = TRUE, stringsAsFactors = FALSE)
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
Mark Wheldon
}
