% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator_functions.R
\name{name}
\alias{name}
\title{Country or region name}
\usage{
name(code)
}
\arguments{
\item{code}{ISO code for the country or UN code for region.}
}
\value{
A vector the same length as \code{code} containing the
    name(s) as character.
}
\description{
Returns the name associated with \code{code}. If a region
is given, all names associated will be returned in a vector.
}
\examples{

## Countries
name(c(250, 4))

## Regions
name(900)

}
\seealso{
Other translator functions: 
\code{\link{code}()},
\code{\link{country_codes}()},
\code{\link{country_names}()},
\code{\link{reg_code}()},
\code{\link{reg_name}()}
}
\author{
Mark Wheldon
}
\concept{translator functions}
