% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator_functions.R
\name{reg_name}
\alias{reg_name}
\title{Region name associated with a country or region}
\usage{
reg_name(
  x,
  level = c("1", "2", "other"),
  family = c("M49", "SDG", "WB_inc", "Dev")
)
}
\arguments{
\item{x}{\emph{Region} identifier. Interpreted as \dQuote{code}
if \code{is.numeric(x)} and \dQuote{name} if
\code{is.character(x)}.}

\item{level}{Level of region. Higher levels are nested in lower
levels. E.g., \dQuote{Africa} is level \dQuote{1},
\dQuote{Eastern Africa} is level \dQuote{2}. Converted to
character if supplied as numeric. Level \dQuote{other} is for
regions that are not part of an official classification, such
as \dQuote{Europe} in the \dQuote{SDG} family and
\dQuote{middle income} in the \dQuote{WB_inc} family.}

\item{family}{Family to which the region referenced by \code{code}
belongs.}
}
\value{
Region name as character
}
\description{
Returns the region name associated with a given country code.
}
\details{
If a region name is supplied it will just be returned.
}
\examples{

## Default family is "M49"
reg_name(250)

reg_name(250, family = "SDG")

## Level can be supplied as numeric for convenience
reg_name("France", 2, family = "SDG")

}
\seealso{
Other translator functions: 
\code{\link{code}()},
\code{\link{country_codes}()},
\code{\link{country_names}()},
\code{\link{name}},
\code{\link{reg_code}()}
}
\author{
Mark Wheldon
}
\concept{translator functions}
