% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{reg_table}
\alias{reg_table}
\title{Create table of countries and regions}
\usage{
reg_table(
  level = c("", "1", "2"),
  family = c("M49", "SDG", "WB_inc", "Dev"),
  include_codes = TRUE,
  include_names = TRUE,
  stringsAsFactors = FALSE
)
}
\arguments{
\item{level}{Level of region. Higher levels are nested in lower
levels. E.g., \dQuote{Africa} is level \dQuote{1},
\dQuote{Eastern Africa} is level \dQuote{2}. Converted to
character if supplied as numeric. The \sQuote{""}, or
\code{NULL}, can be given to request all levels.}

\item{family}{Family to which the region referenced by \code{code}
belongs.}

\item{include_codes}{Logial. Include country and region codes in
the result? You must set at least one of \code{include_codes}
and \code{include_names} to \code{TRUE}.}

\item{include_names}{Logical. Include country and region names in
the result?}

\item{stringsAsFactors}{Passed to \code{\link{data.frame}}. Note
the default is \code{FALSE}.}
}
\value{
A data frame with country codes and names (if requested)
    and additional columns for the region codes and names.
}
\description{
Creates a data frame with columns for country names and codes (if
requested) and corresponding regions (as requested).
}
\details{
Unlike \code{\link{reg_code}} and \code{\link{reg_name}}, if
\code{level} is not specified both are returned.
}
\examples{
reg_table(level = 1, family = M49)

}
\seealso{
Other Table functions: 
\code{\link{country_table}()}
}
\author{
Mark Wheldon
}
\concept{Table functions}
