% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{shortern_reg_names}
\alias{shortern_reg_names}
\title{Shorten region names}
\usage{
shortern_reg_names(names)
}
\arguments{
\item{names}{Character vector of names to shorten.}
}
\value{
Character vector of shortened region names.
}
\description{
Shorten region names to help with plotting, for example.
}
\author{
Mark Wheldon
}
