% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{sub_sdg_names}
\alias{sub_sdg_names}
\title{SDG region name variants}
\usage{
sub_sdg_names(x)
}
\arguments{
\item{x}{Character vector containing names to substitute.}
}
\value{
Character vector after substitution.
}
\description{
Substitutes SDG region names for a variant.
}
\details{
This currently concerns only \dQuote{Australia/New Zealand} and
\dQuote{Oceania}.
}
\author{
Mark Wheldon
}
