% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unreg.R
\docType{package}
\name{unreg}
\alias{unreg}
\alias{unreg-package}
\title{unreg: Utilities for accessing country ISO codes and UN geographic regions}
\description{
unreg: Utilities for accessing country ISO codes and UN geographic regions
}
