% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator_functions.R
\name{code}
\alias{code}
\title{Country or region code}
\usage{
code(name, family = c("M49", "SDG"))
}
\arguments{
\item{name}{Name of country or region (case insensitive).}
}
\value{
A vector the same length as \code{code} containing the ISO
    code(s) as numeric.
}
\description{
Returns the ISO code(s) associated with \code{name}. If a region
is given, all codes associated will be returned in a vector.
}
\seealso{
Other translator functions: \code{\link{country_codes}},
  \code{\link{country_names}}, \code{\link{name}},
  \code{\link{reg_code}}, \code{\link{reg_name}}
}
\author{
Mark Wheldon
}
\concept{translator functions}
